Extra safeties in copy script
--HG--
branch : develop

+++ xnat/scripts/copy_project.py
@@ -64,15 +64,18 @@ class XNATProjectCopier:
             raise ValueError('Could not find directory for downloaded resource!')
 
         # Upload entire resource directory
-        dest_resource.upload_dir(resource_path, method='tgz_file')
+        dest_resource.upload_dir(resource_path, method='tgz_memory')
 
     def copy_resources(self, source_object, dest_object, prefix=''):
-        for resource_id, source_resource in source_object.resources.items():
+        for source_resource in source_object.resources.values():
+            resource_id = source_resource.label
             if resource_id in dest_object.resources:
+                print('{}Skipping resource {}'.format(prefix, resource_id))
                 # Already there, do not copy
                 continue
 
             # Create a resources of the same xsitype
+            print('{}Copying resource {}'.format(prefix, resource_id))
             dest_class = self.dest_xnat.XNAT_CLASS_LOOKUP[source_resource.__xsi_type__]
             dest_resource = dest_class(
                 parent=dest_object,
@@ -87,6 +90,21 @@ class XNATProjectCopier:
     def copy_experiment(self, source_experiment, dest_experiment):
         self.copy_fields(source_experiment, dest_experiment, prefix='    ')
         self.copy_resources(source_experiment, dest_experiment, prefix='    ')
+
+        for scan in source_experiment.scans.values():
+            print('    copying scan {} / {}'.format(scan.id, scan.type))
+            if scan.id not in dest_experiment.scans:
+                dest_class = self.dest_xnat.XNAT_CLASS_LOOKUP.get(scan.__xsi_type__)
+                if dest_class is None:
+                    print('     [WARNING] {} class not found on destination server, skipping'.format(source_experiment.__xsi_type__))
+                    continue
+
+                print('     creating scan {}'.format(scan.id))
+                dest_scan = dest_class(parent=dest_experiment, id=scan.id, type=scan.type)
+            else:
+                dest_scan = dest_experiment.scans[scan.id]
+
+            self.copy_resources(scan, dest_scan, prefix='     ')
 
         for assessor_id, source_assessor in source_experiment.assessors.items():
             # Create an assessor of the same xsitype
@@ -108,20 +126,42 @@ class XNATProjectCopier:
         for source_experiment in source_subject.experiments.values():
             print('  copying experiment  {}'.format(source_experiment.label))
 
-            if len(source_experiment.scans) > 0:
+            if hasattr(source_experiment, 'scans') and len(source_experiment.scans) > 0:
                 print('    copying data')
                 temp_file = os.path.join(self.temp_dir, source_experiment.label + '.zip')
-                source_experiment.download(temp_file, verbose=False)
-                dest_experiment = self.dest_xnat.services.import_(
-                    temp_file,
-                    project=self.dest_project.id,
-                    subject=source_experiment.subject.label,
-                    experiment=source_experiment.label
-                )
-                os.remove(temp_file)
+                try:
+                    source_experiment.download(temp_file, verbose=False)
+                    try:
+                        dest_experiment = self.dest_xnat.services.import_(
+                            temp_file,
+                            project=self.dest_project.id,
+                            subject=source_experiment.subject.label,
+                            experiment=source_experiment.label
+                        )
+                    except xnat.exceptions.XNATUploadError as exception:
+                        print('    [ WARNING] Experiment did not include parsable dicom files, creating empty experiment')
+                        if 'not include parseable files' in exception.args[0]:
+                            dest_class = self.dest_xnat.XNAT_CLASS_LOOKUP.get(source_experiment.__xsi_type__)
+                            if dest_class is None:
+                                print('    [WARNING] {} class not found on destination server, skipping'.format(source_experiment.__xsi_type__))
+                                continue
+
+                            dest_experiment = dest_class(parent=dest_subject, label=source_experiment.label)
+                        else:
+                            raise
+                finally:
+                    try:
+                        os.remove(temp_file)
+                    except:
+                        pass  # Allowed
+
             else:
                 print('    creating empty experiment')
-                dest_class = self.dest_xnat.XNAT_CLASS_LOOKUP[source_experiment.__xsi_type__]
+                dest_class = self.dest_xnat.XNAT_CLASS_LOOKUP.get(source_experiment.__xsi_type__)
+                if dest_class is None:
+                    print('    [WARNING] {} class not found on destination server, skipping'.format(source_experiment.__xsi_type__))
+                    continue
+
                 dest_experiment = dest_class(parent=dest_subject, label=source_experiment.label)
             self.copy_experiment(source_experiment, dest_experiment)
 
@@ -129,8 +169,10 @@ class XNATProjectCopier:
         source_project = self.source_project
         dest_project = self.dest_project
 
+        print('Copying fields')
         self.copy_fields(source_project, dest_project)
-        self.copy_resources(source_project, dest_project)
+        print('Copying resources')
+        self.copy_resources(source_project, dest_project, prefix='  ')
 
         for subject_id, source_subject in self.source_project.subjects.items():
             print('copying subject {}'.format(source_subject.label))
@@ -145,8 +187,8 @@ def main():
     parser = argparse.ArgumentParser(description='Copy Xnat projects')
     parser.add_argument('--source-host', type=six.text_type, required=True, help='source XNAT url')
     parser.add_argument('--source-project', type=six.text_type, required=True, help='source XNAT project')
-    parser.add_argument('--dest-host',  type=six.text_type, required=True, help='destination XNAT url')
-    parser.add_argument('--dest-project',  type=six.text_type, required=True, help='destination XNAT project')
+    parser.add_argument('--dest-host', type=six.text_type, required=True, help='destination XNAT url')
+    parser.add_argument('--dest-project', type=six.text_type, required=True, help='destination XNAT project')
     args = parser.parse_args()
 
     with xnat.connect(args.source_host) as source_xnat, xnat.connect(args.dest_host) as dest_xnat:
